Added the second section with the 4 squares.
I decided to make these squares paint samples so I didn't
need to find a bunch of images that fit as perfect squares.
I also replaced the text with some goofy text describing a store
that sells paint and cats. One problem that I'm having at the
moment is that I can't get the image to shrink. I'm not sure why.
I think it might be related to containers not shrinking.